+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Bug fixes
+
+* [#9612](https://github.com/rubocop/rubocop/pull/9612): Allow parentheses in safe navigation calls for `Style/MethodCallWithArgsParentheses` `EnforcedStyle: omit_parentheses`. ([@gsamokovarov][])
+
 ## 1.11.0 (2021-03-01)
 
 ### New features